requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.26.0
google-generativeai==0.3.2
requests==2.31.0
